notebooks/requirements.txt
numpy
matplotlib
pytesseract
pillow
nltk
pdf2image
pyyaml
